+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes starting with the version 0.6.9 are documented here.
 
 ## [0.9.0]
 - Update blas-src dependency
+- Implement inverse of lower/upper regularized incomplete gamma function
+- Improve accuracy of of the quantile function of the Chi-square distribution
 
 ## [0.8.1]
 - README corrections